Added text box
Added a very basic text box for debugging and would probably be used in future for player name input etc and cheat input (we will probably just rename the debug stuff to cheats).

Added basic message list which just shows up to 10 recently used debug commands for now. Will be used for chat later on once multiplayer is added.

Added measuring of individual characters in the renderer which needed a the GDI+ to have a working IFont allocation (it just returns a wrapper around a font object) but in doing this, we prevent less hits on memory per call as OpenGL would of just called AllocateFont for every call of GetCharWiidth.