Added "hot loading"
Added the ability to hot-load files potentially containing data about game entities or settings. These "variables" can be changed from a file and the effects are almost immidiate.

I yet to create documentation for the syntax that the hotloader uses on files loaded in but it has the following features:
1) Classes (to keep variables organized and given a context)
2) Variables can either be "static" or "const". "static" means once the variable is loaded, any future adjustments to the expression will not effect the one in memory. "const" means the engine cannot make changes to the value.
3) Expressions can reference other variables in other classes
4) Comments are supported ("#" character)

Todo:
1) Add a better syntax error log instead of just throwing an exception
2) Support negative integers